+++ build.gradle.kts
@@ -2,8 +2,17 @@ plugins {
     id("org.rivierarobotics.gradlerioredux") version "0.9.7"
 }
 
+afterEvaluate {
+    tasks.withType<Jar>(){
+    duplicatesStrategy = DuplicatesStrategy.INCLUDE
+}}
+
 gradleRioRedux {
-    robotClass = "org.rivierarobotics.robot.Robot"
-    teamNumber = 5818
+    robotClass = "org.octobots.robot.Robot"
+    teamNumber = 9084
+
     applyGradleRioConfiguration()
 }
+
+
+
