+++ gradle.properties
@@ -1,4 +1,4 @@
-group=org.rivierarobotics
-url=https://github.com/Team5818
-organization=Riviera Robotics
+group=org.teamoctobots
+url=https://github.com/Octobots9084
+organization=Octobots
 version=0.1.0-SNAPSHOT
